user bond info moved into account substate (#381)
* user bond info moved into accout substate

* got rid of extra line

* moved bondasset back to bondslice

* pre-develop change
