+++ src/helpers/index.js
@@ -240,3 +240,9 @@ export function setAll(state, properties) {
     state[key] = properties[key];
   });
 }
+
+export const setBondState = (state, payload) => {
+  const bond = payload.bond;
+  const newState = { ...state[bond], ...payload };
+  state[bond] = newState;
+};
